Change 'size' param of avifIOReadFunc to size_t
Change the 'size' parameter of avifIOReadFunc and avifIOWriteFunc to
size_t, because it represents the size of a memory buffer. Only the size
and offset of a file need to be a 64-bit type (in order to support large
files > 2 GB).

Fix https://github.com/AOMediaCodec/libavif/issues/357

+++ tests/aviftest.c
@@ -261,9 +261,9 @@ typedef struct avifIOTestReader
     size_t availableBytes;
 } avifIOTestReader;
 
-static avifResult avifIOTestReaderRead(struct avifIO * io, uint32_t readFlags, uint64_t offset, uint64_t size, avifROData * out)
+static avifResult avifIOTestReaderRead(struct avifIO * io, uint32_t readFlags, uint64_t offset, size_t size, avifROData * out)
 {
-    // printf("avifIOTestReaderRead offset %zu size %zu\n", offset, size);
+    // printf("avifIOTestReaderRead offset %" PRIu64 " size %zu\n", offset, size);
 
     (void)readFlags;
